Extract `printArguments`
